+++ docs/src/markdown/about/changelog.md
@@ -7,6 +7,9 @@
     `[[:xdigit:]]`, and  `[[:punct:]]`. To explicitly use standard Unicode rules for these compatibility properties, use
     the Unicode property form instead: `[\p{Alnum}]`, `[\p{Digit}]`, `[\p{Punct}]`, or `[\p{XDigit}]`. This has changed
     to ensure no confusion for users expecting compatible POSIX style character class properties.
+-   **FIX**: Scoped ASCII/Unicode flags (`(?a:pattern)`/`(?u:pattern)`) should be respected for Unicode properties in
+    `bre`.
+-   **FIX**: Fix issues with disabled scoped flags.
 
 ## 5.9